+++ lib/trufi_app.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:package_info/package_info.dart';
-import 'package:trufi_core/blocs/app_review_bloc.dart';
+import 'package:trufi_core/blocs/app_review_cubit.dart';
 import 'package:trufi_core/blocs/home_page_bloc.dart';
 import 'package:trufi_core/blocs/request_manager_bloc.dart';
 import 'package:trufi_core/l10n/material_localization_qu.dart';
@@ -21,7 +21,7 @@ import './blocs/favorite_locations_bloc.dart';
 import './blocs/history_locations_bloc.dart';
 import './blocs/location_provider_bloc.dart';
 import './blocs/location_search_bloc.dart';
-import './blocs/preferences_bloc.dart';
+import './blocs/preferences_cubit.dart';
 import './blocs/saved_places_bloc.dart';
 import './pages/about.dart';
 import './pages/feedback.dart';
