+++ pubspec.yaml
@@ -10,7 +10,6 @@ environment:
 
 dependencies:
   app_review: ^2.1.1+1
-  bloc_test: ^8.0.0
   cupertino_icons: ^1.0.2
   diff_match_patch: 0.3.0
   equatable: ^2.0.0
@@ -27,7 +26,6 @@ dependencies:
   google_api_availability: ^3.0.1
   http: 0.13.1
   intl: 0.17.0
-  mockito: 4.1.4
   package_info: 2.0.0
   path_provider: ^2.0.1
   permission_handler: ^6.1.1
@@ -40,11 +38,13 @@ dependencies:
   uuid: 2.2.2
 
 dev_dependencies:
+  bloc_test: ^8.0.0
   flutter_driver:
     sdk: flutter
   flutter_test:
     sdk: flutter
   network_image_mock: 1.1.0
+  mockito: 4.1.4
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
