+++ test/blocs/preferences_cubit_test.dart
@@ -1,6 +1,6 @@
 import 'package:bloc_test/bloc_test.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:trufi_core/blocs/preferences_bloc.dart';
+import 'package:trufi_core/blocs/preferences_cubit.dart';
 import 'package:trufi_core/models/preferences.dart';
 
 import '../mocks/local_repository_mock.dart';
@@ -8,7 +8,7 @@ import '../mocks/local_repository_mock.dart';
 void main() {
   group("PreferencesCubit", () {
     blocTest(
-      "should add another Preference to the state",
+      "updateMapType emits [Preference, Preference]",
       build: () => PreferencesCubit(MockLocalRepository()),
       act: (PreferencesCubit cubit) => cubit.updateMapType("TestMapTyle"),
       expect: () => [
